requirements.txt
fastapi
uvicorn[standard]
sqlalchemy
sqlmodel
pydantic
python-jose[cryptography]
passlib[bcrypt]
python-multipart
email-validator
alembic
pymysql
